Add downloading from url payload type
Summary:
Adds a payload type for downloading from a url. This can take advantage of `bsdtar`s archive detection, which means that it's easy to bridge a data stream task to the same way we'd stream data over grpc.

This does require that the companion-side has access to the url, but failure still occurs if the extract fails for any reason, for instance a garbage input, or 404

Reviewed By: zeyadsalloum

Differential Revision: D15117099

fbshipit-source-id: 97d6b6f03783c4d76f2a85a2bbcdb8f5e969ed5d