Add client-side support for installing an app via a url
Summary:
This adds support for an app install via a URL.

The companion now trivially supports installs from a zip or an ipa.

We can also rely on the behaviour of `bsdtar` to explode if something bad
happens in the download on the remote side.

There should eventually be a local-backoff, but for the time being this still
works.

One very nice attribute here is that `bsdtar` also does *incremental
extraction* of IPAs, meaning that the unzipping occurs whilst the ipa is being
downloaded. This means that the extract and download run in parallel with each
other.

Reviewed By: zeyadsalloum

Differential Revision: D15118655

fbshipit-source-id: b5b5f8115817c06c7c89836a1b6518ef72aeec5d