Add cancel_wrapper for streams
Summary:
Currently, we use the async iterator on a stream, and then check that
the stop future is set.

However, this means that the check is performed when the next yield happens. We
want to interrupt any active stream and cancel when stopped. This can also then
be repeated against other commands

Reviewed By: c-ryan747

Differential Revision: D15118400

fbshipit-source-id: 8ab4b34c984e7862ef06433687250286087798aa