Add @autoreleasepool to all handlers
Summary:
This clearly isn't the best solution right now, but this at least stops the bleeding i.e. leaks of NSObjects that are in an autoreleasepool in a background thread.

This all came up as I was seeing that `FBSimulatorAgentOperation` wasn't deallocated after `list-apps`. This was because the final `release` did not occur, as it was held onto by an `autoreleasepool` in a background thread.

`autoreleasepool`s can be appended to just fine in this background thread, but the releasing of the pool's objects never occurs, because there's nowhere that the pool drain is executed. Since gRPC executes this in a background thread, we have to create a pool for these root threads.

This isn't ideal in terms of an ideal approach. Ideally, we'd have a wrapper class that can inject the pool in a call, but my C++ template skills aren't there yet. We could also use macros, but that doesn't make things any clearer.

Reviewed By: zeyadsalloum

Differential Revision: D15278541

fbshipit-source-id: 3beed500f274c28394f8c015563fbd5580e6728a