remove streaming from default_daemon
Summary:
the default daemon implementation acted as a forwarder. so for all the cases where the daemon was just forwarding the request to the companion the default implementation was enough and no daemon method needed to be implemented in the respective command file.
The problem was that the default daemon implementation was handling steams and unary methods. and it was calling recv_message more than once on all of them. this is fine in the streaming but it was leading to weird behaviour when it was a unary.

given that the majority of the commands were unary. It was simpler to make the default method just handle unary for simplicity and the streaming commands should just implement a custom daemon method which i'm doing in the coming diffs.

Reviewed By: lawrencelomax

Differential Revision: D15279224

fbshipit-source-id: 3ea63d6b371b8038d36acc97b53caf494c68fafb