Prevent FD leakage on Agent Operation termination
Summary:
Mostly in the comments, but we need to take account of the fact that CoreSimulator doesn't close stdout/stderr of a spawned process for us.

I discovered this when noticing that there's a 4s delay in detaching stdout/stderr of a pipe after list-apps. This further revealed that the `-[FBProcessOutput_Consumer detach]` implementation that waited for it's `FBFileReader` to close, would always timeout internally. This was the 4s delay.

Since spawned tasks don't have this problem, this reveals that it's something exclusively in agent operations.

Reviewed By: zeyadsalloum

Differential Revision: D15279211

fbshipit-source-id: 1bc650a9ecda98bc78f0833034b4918b7c75ef19