Smooth over API cracks in NSTask
Summary:
Instead of having three separate properties:
- isRunning
- exitCode
- terminationStatus

We can have one, a `FBFuture<NSNumber *> *`. If the future is running then isRunning = YES, if completed it's done.

This makes it easier to add other implementations and removes this duplication. It also prevents us from ever from the sharp edge of raising an exception if `terminationStatus` is called when `isRunning` is false

Reviewed By: c-ryan747

Differential Revision: D15318958

fbshipit-source-id: 15cb45f6d178f1f887aa46e3f45a1ac43198ffac