Fix xctest_result_bundle output for xctest command
Summary:
Noticed that the result bundle was empty for xctest over grpc.
Reason is that delta.resultBundlePath is never used when creating a response in FBIDBServiceHandler::xctest_run

Fixed by using the value from the delta when creating a response. However, to avoid sending the bundle over for every delta, only do it when command status is Terminated.

Reviewed By: lawrencelomax

Differential Revision: D15321417

fbshipit-source-id: 0e0842b2c6c203f5ae9f92b656d72751a492935d