NSBundle -> FBBundleDescriptor
Summary:
Uses `NSBundle` APIs instead of doing this all for ourselves.

We currently do all the parsing and rules around this ourselves, which is massively superflous since `NSBundle` has many of the APIs that we care about.

This also opens the door for allowing bundles without identifiers, which is very common in xctest usage

Reviewed By: zeyadsalloum

Differential Revision: D15354895

fbshipit-source-id: 874a10201dcf052282964a2d1132679317804cc9