Remove DLDeviceListener
Summary:
This removes the need for a "Device Manager" specialized to `DLDevice`. This also makes the threading a lot easier to grok and the whole flow is simplified.

This means we short-circuit using DeviceLink to do the instantiation of `DLDevice` structs. Fortunately, we've already documented what these structs look like. For reference on how they are created, see `AMDeviceNotification` in `DeviceLink.framework`

Doing the instantiation ourselves means a number of things:
1) No need for udid `NSNotifications`
2) `FBDLDevice` becomes part of the `FBDevice` object graph
3) `FBDLDevice` is lazily instantiated. This is important as we only need it for screenshots.
4) Far less indirection with the manager managing devices and connections

Reviewed By: zeyadsalloum

Differential Revision: D15372185

fbshipit-source-id: 08edea3201bbe274d6436cfb125ea61a794a87a7