Fix confusing nominal types for test bundles (D15071258 Redux)
Summary:
This was reverted in a stack, there's been some rebasing since but the original summary is

These are incorrectly specified to be the class `FBApplicationBundle`. There's actually no difference in the implementation, just a subclass that doesn't do anything other than whenit used to have a clever constructor.

Now that's on the base class, we can ensure that this isn't confused as an `FBApplicationBundle`

Reviewed By: zeyadsalloum

Differential Revision: D15371459

fbshipit-source-id: b306a02c32d255d4469c5ce0e9d133e40da54c57