expand app bundle storage location
Summary:
`instrumental` needs to know the location of the app for call stack symbolication. `instrumental` also needs dSYM files, but it does not require their locations as long as they are within the scope of Spotlight.

Add a method to FBApplicationBundleStorage to interpolate a pre-defined const string with the actual path of the app bundle.

Reviewed By: lawrencelomax

Differential Revision: D15332525

fbshipit-source-id: 353d24af31236c22b4157bf0e6a19f0103fc13e4