Use convenience for empty context wrapper
Summary:
These are equivalent, we just need this to ensure that the we conform to the same type. This is a "context that does no teardown"

The reason for this existing is that we don't want to remove individual files when the parent is passed local files that should be unmodified.

There's even an optimization here which is that the teadown is never needed and never added to the teardown stack

Reviewed By: zeyadsalloum

Differential Revision: D15431327

fbshipit-source-id: 6ea62b5c68a30489141f28772c4843ec16abcda1