use file storage instead of bundle storage for dSYM files
Summary:
The dSYM files are bundles but they do not have all the required fields for `FBBundleDescriptor`, so use `installBundle` for a dSYM file would fail. Additionally, we usually retrieve dSYM files from URL, which points to a folder that contains multiple dSYM files. I tried to implement installing multiple dSYM files at the same time in D15430231, but the install would fail for the same reason as installing a single dSYM file. `FBFileStorage` seems to be an easy solution.

This is one example plist for the dSYM file:
```
<?xml version="1.0" encoding="UTF-8"?>
<!DOCTYPE plist PUBLIC "-//Apple Computer//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
<plist version="1.0">
	<dict>
		<key>CFBundleDevelopmentRegion</key>
		<string>English</string>
		<key>CFBundleIdentifier</key>
		<string>com.apple.xcode.dsym.BroadcastUIExtensionCompressed-iphoneos-arm64Binary_raw#apple-dsym,iphoneos-arm64</string>
		<key>CFBundleInfoDictionaryVersion</key>
		<string>6.0</string>
		<key>CFBundlePackageType</key>
		<string>dSYM</string>
		<key>CFBundleSignature</key>
		<string>????</string>
		<key>CFBundleShortVersionString</key>
		<string>1.0</string>
		<key>CFBundleVersion</key>
		<string>1</string>
	</dict>
</plist>
```

Reviewed By: lawrencelomax

Differential Revision: D15430351

fbshipit-source-id: 765e056f60b8e242759ecebca2d1286400f22312