Don't override `closeOnDealloc` behaviour
Summary:
`closeOnDealloc` is being disobeyed in `FBFileWriter`. This is particularly problematic with sockets, which share a file descriptor on the reading and writing end of the `fd`, and can result in the read end getting interrupted despite as the write end has closed the file descriptor.

This should solve a lot of the `EV_VANSIHSED` problems that exist.

Reviewed By: zeyadsalloum

Differential Revision: D15432072

fbshipit-source-id: ba3e479da106bf8d86afed1066a39f2cf15446e7