Pure AMDevice Implementation of Screenshot
Summary:
This makes a pure `AMDevice` implementation with internal classes.

The intention here is that we can actually get rid of the `DeviceLink` implementation and therefore our dependency on it.

This also means that we don't have to fake `DeviceLink` connections and can more precisely work with connections to these services

Reviewed By: zeyadsalloum

Differential Revision: D15416111

fbshipit-source-id: ee0831d5d7486c5ced4f60078ad8e63efc2db890