Allow teardown to be asynchronous
Summary:
If there are stacked contexts, the order in which the teardown handlers are called is deterministic, but if there's any async work that happens in them these can intefere with one another.

For example in the case that:
- A context is pushed that makes a socket connection
- A context is pushed that reads the socket
- The socket reader is used.
- The socket read context is popped, but the teardown is async
- The socket connection context is popped, but the teardown is not async.

This means that the socket read teardown may not have fully completed by the time that the socket is popped. This can lead to the classic `EV_VANISHED` if the reader is in fact a dispatch one.

Reviewed By: zeyadsalloum

Differential Revision: D15431291

fbshipit-source-id: 9c591d7284a8bc314386682b2023047689df2958