Explicit closing of file descriptors in FBFileWriter
Summary:
This is a common pain point and one I've put off fixing for far too long.

Treating `dealloc` as a way of tearing down resources is a fools-errand and in many respects similar to the perils of doing this in Java/Python with object finalizers. There's well-understood reasons for making resource cleanup explicit in these cases. In GC'd languages in Python, contextmanagers exist to make closing of resources such as file descriptors very explicit (but convenient!) and objects wrapping file descriptors will warn and possibly leak in the event the resource has not been closed before finalization.

Whilst Objective-C isn't Garbage Collected and ARC is mostly more deterministic, the end result of using `dealloc` for closing resources that isn't just memory can have exactly the same problems:
- If the `NSFileHandle` is "leaked" or abandoned on the heap somewhere, the resource will never close. Leaking memory is clearly bad, but it's better if these leaks don't have a downstream affect on resource leakage. The problem of memory leaks becomes much more of a problem when it results in resource leakage, or even worse causing hangs because piping data between processes hangs forever.
- `FBFileWriter` is used for different cases where we may more may not want to close the file descriptor when we're done writing. For `AMDevice` services, we already have closing semantics in the context managers and don't want to double-close. For writing to a file path, we want to close as soon as we're done. Currently, this is encoded in the `closeOnDealloc` behaviour of `NSFileHandle` but this moves the resource management problem to a reference-counting one.
- There's no way outside of private APIs in `NSFileHandle` to query if `closeOnDealloc` is set and then explicitly calling this. Instead we can encode that knowledge in the writer itself.
- There's no guarantee that the setting `self.fileHandle` to `nil` will immediately result in the closing of the file, should `closeOnDealloc` be set. This may be because the `NSFileHandle` is legitimately referenced elsewhere, so it shouldn't be closed. Or it could be held onto in a autoreleasepool, in which case the closing of the file will occur much later than desired. All of this adds non-determinism.

Now, all of this is completely explicit and we'll do the same on the reader side. I also recall having a conversation about this with clayborg and now realize just how right he was!

Reviewed By: zeyadsalloum

Differential Revision: D15450350

fbshipit-source-id: 46ccb07953fadf2f668db5707cc673cc2e3abfb1