Add sizing of transfer
Summary:
This is helpful for determining the number of bytes sent over the wire when installing a stream of data.

This means that we add a new field to the event reporter, and then conditionally add to this if the first argument of the method has a declared `bytesTransferred` method declared. At the moment we will only see this logged in `install_stream` and `install_xctest_stream` but over time this can be expanded by manipulating and moving protocols around.

Reviewed By: c-ryan747

Differential Revision: D15555022

fbshipit-source-id: d3fd4fb56ded198fc139c7e4c412574e897787bf