Fix detach race in FBProcessOutput
Summary:
There is a race where nulling out a pipe will cause the read end to be closed whilst there are active readers.

This ensures that the write end is closed on detachment, but that the read end is still left open or ready for the reciever.

Eventually we should make this clearer by having separate detachment of the read and the write end, but for now this at least prevents this race:
1) Process Dies
2) Detach is called
3) Write end is closed
4) Pipe is deallocated, but the read end is referenced as a file descriptor and therefore does not retain the file descriptor
5) EV_VANISHED on a read.

Reviewed By: c-ryan747

Differential Revision: D15576202

fbshipit-source-id: 3c40687e68ba0c2bdcd9364849d546a9d734a642