Fix Consumer Output not closing write end of the file handle
Summary:
The detachment wasn't calling super, this means that the write end of the handle isn't being closed.

For an `NSTask` it is the case that this happens automatically, however for other cases this isn't true. As a result we need to ensure the file is closed. If the file is closed twice then this is a no-op

Reviewed By: c-ryan747

Differential Revision: D15576269

fbshipit-source-id: ea2f0b85d4750b1cd962b9181094ad9caf3d8ce0