Close read end of consumer
Summary:
There's no concern that we have with sockets where the read and the write end are the same fd. A pipe will produce two file descriptors, instead of the same descriptor.

Therefore we don't need to worry about closing the read end of the pipe when we're done. The `NSPipe` may still be retained in this state, but this is fine since both the read and write ends will be closed when all channeling is done. It's just the deallocation of the pipe object wrapper that will happen on dealloc

Reviewed By: c-ryan747

Differential Revision: D15576286

fbshipit-source-id: 07c266d08c4cc4ff9e0a2f9126426e28a1412539