Remove erroneous /bin/sleep responds to SIGINT tests
Summary:
This is actually innacurate. `NSTask` does something different in `-[NSTask interrupt]` other than sending a `SIGINT`.

This means this isn't as simple as sending a `SIGINT`. As such these tests are erroneous as they don't just `SIGINT`.

If I find a better test for a process that responds well to a `SIGINT`, these tests can be resurrected

Reviewed By: c-ryan747

Differential Revision: D15600756

fbshipit-source-id: ebf07b9dfac40b5fe017234038e1cbf89ce0d649