Replace NSTask implementation with posix_spawn (D15600757 Redux)
Summary:
The envp was erroneously implemented. The correct form is an array of `key=value` for each environment variable, instead of `[key, value]`.

This comes from the documentation for `execve`

```
envp is an
array of strings, conventionally of the form key=value, which are
passed as environment to the new program.  The argv and envp arrays
must each include a null pointer at the end of the array.
```

Reviewed By: c-ryan747

Differential Revision: D15621497

fbshipit-source-id: 3e419a80e8091b40b6f596d1cff893cc75af4085