Replace NSTask implementation with posix_spawn (D15621497 Redux)
Summary:
This was a fun one to diagnose. The implementation of `NSTask` sets the signal mask and defaults in the child process with `posix_spawnattr_setsigmask` and `posix_spawnattr_setsigdefault`.

Additionally, these need to be flagged in `posix_spawnattr_setflags`.

The symptom was that if a launched subprocess launched a child proces, then the child process may not be notified of the completion of it's child. This is because (depending on the calling process), the default disposition of `SIGCHILD` may not be correct. As a result the child process cannot reasonably understand when a child dies, unless it re-sets the disposition of `SIGCHILD`.

Reviewed By: c-ryan747

Differential Revision: D15712821

fbshipit-source-id: f1629772ac4149b4bbfec5cef666621f44b2f87d