No NSPipe or NSFileHandle in FBProcessStream
Summary:
We can instead use the raw fd based approach and also include whether this is input or output.

This also fixes a bug whereby a file would always be created before opening a file, this caused problems when using `/dev/null` as a file path, as the creation was *mandatory*. Now with the `O_CREAT` flag, this is *optional*, so if the file path exists already (as is the case with `/dev/null`) we won't bail out.

In fact it is also possible to use `-1` as a file descriptor instead. Indeed this is the approach that `NSNullFileHandle` takes, so that there's no need to open additional file desciptors to nothing. This is further optimization that we can make in the future.

Reviewed By: c-ryan747

Differential Revision: D15621942

fbshipit-source-id: 7b4522a032ec8888bd8449915ac0f0d7b9f5f52a