Replace space with underscore in trace name
Summary:
It turns out that gRPC idb instruments fails to transfer traces from idb companion host to idb client host if the trace name contains space. idb uses the template name to construct the default trace path if one is not provided. If the template name contains space, the file name will contains space.

Replace space with underscore for default file names. Change on the companoin side in this diff is not necessary but it's handy to manually deal with file names without space.

Reviewed By: lawrencelomax

Differential Revision: D15685571

fbshipit-source-id: cf9fdd0f5abb971f69a552d6742430932eaf2159