Do not signal a read has completed on "done"
Summary:
There's another race condition related to file reading that's not yet been plugged. This relates to the difference between a "done" on `dispatch_io_read` and the `dispatch_io_close` callback:

1) A read operation has finished on a file descriptor.
2) The caller is observing `finishedReading` which now resolves. The caller expects that once this is resolved the file descriptor is free to use in any way.
3) The dispatch channel is not truly relinquished until the callback of `dispatch_io_close` is called. This may race with the above.
4) The caller then closes the file descriptor, causing an `EV_VANISHED`.

This diff just ensures that any futures that are vended here correspond to the `dispatch_io_close` callback instead of the `dispatch_io_read` (done) callback. When profiling this it's certainly the case that the "done" occurs before the relinquished control callback.

In fact the `dispatch_io_create` API makes it clear that the control relinquishment only occurs in this particular callback:

> The block to enqueue when the system relinquishes control of the channel’s file descriptor.

There is no such mention in the `dispatch_io_read` callback:

> If the done parameter is set to YES, it means the read operation is complete and the handler will not be submitted again
> If the handler is submitted with the done parameter set to YES, an empty data object, and an error code of 0, it means that the channel reached the end of the file

This means that these two callbacks may be related, but fd control relinquishment will occur *after* the read operation is done. This does make sense, there can be many read operations per channel but only one fd per channel. It stands to reason that we should instead only consider the relinquishment as passing control back to the consumer

Reviewed By: c-ryan747

Differential Revision: D15740960

fbshipit-source-id: 8c8191a7578f57bd10485cef5554b0c8c74ee72f