Add error handling to FBProcessStream to NSOutputStream
Summary:
This is something that should be overridden in sublcasses of `NSOutputStream`.

We can now handle this at the lower level of the `FBTask` and confirm that there has been no error involved in the reading or writing of a stream.

This is preferrable to these things failing implictly on a broken pipe and means we'll get a much more localized failure message in these cases

Reviewed By: zeyadsalloum

Differential Revision: D15621982

fbshipit-source-id: 98fcb8a96904bf8ba7c76c6aeaaf95f7b9d1abe9