Add custom idb logger implementation
Summary:
This will allow consumers to see what is going on in the companion, which is handy for debugging on the client side.

The first step is to create a custom subclass, then we can add additional behaviour into the handler.

We need to override globally so that we can inject the loggers into any derived sub-loggers. Since we are a subclass, this is where we can inject the additional loggers

Reviewed By: c-ryan747

Differential Revision: D15803176

fbshipit-source-id: a0eabf26c248074647d66ca590a006a6b6711e95