Add FBInstalledArtifact
Summary:
The current values returned by FBIDBStorageManager are just strings. Additionally, the value isn't a path or bundle id, it's some unique name to identify the artifact.

Instead, wrap this into an object, which will allow us to add additional values such as UUIDs that represent a unique identifier associated with a binary (if present).

Reviewed By: c-ryan747

Differential Revision: D15822522

fbshipit-source-id: 92acfaeeed4909867672289590ec9f6cc3e99328