Add LC_UUID to Binary Descriptor
Summary:
Adds parsing of the LC_UUID to the binary descriptor.

This is possible with otool, but since we already have parsing in place, we can just extend it with some more functionality that will allow us to dive into the commands and pull out the uuid. This will break out if we don't manage to get to all of the load commands.

Reviewed By: c-ryan747

Differential Revision: D15821788

fbshipit-source-id: f43ba24c4cbdc5208189db8b829ca9a52b73f253