Return UUID and Name in grpc
Summary:
If the UUID is parsed from a binary it can also be used in place of the name of the bundle.

We only currently expose the "bundle_id" but this is actually incorrect as it will fallback to some other name if not present. In future we can re-name the proto so that it's clear that this is a generic name identifier and an (optional) UUID that can be derived from the binary itself.

Reviewed By: zeyadsalloum

Differential Revision: D15822521

fbshipit-source-id: abcab24d457b02dca29cd8602857a048e715d5f4