Support Relocations and Codesigns via rpath detection
Summary:
Adds the ability to relocate binaries when they contain Xcode paths. This is mainly for the Detox flow, where there are rpaths located inside the Framework that points to a burnt-in Xcode path that may not be present on the remote end.

Since we still need this rpath, we need to change these rpaths and also ensure that they reference the new Xcode path.

Reviewed By: zeyadsalloum

Differential Revision: D15855754

fbshipit-source-id: 6237e3fe584ff14b7ebca2b8e3ad870341422f6a