Add stderr logging for MobileDevice.framework
Summary:
I'll extend this in the future, but this will mean that we can at the very least get full logging when `mobdevlog` is called inside `MobileDevice.framework`.

This isn't as easy as the `SimLogSetHandler` in `FBSimulatorControl` as there's no easy way for us to set the `gAsl` value that is used as the asl client. Also asl!

To get this working we instead hook the `asl_open` function used in `mobdevlog`, which is statically initialized. THere's no string label for this logging facility, but we can instead know that the `asl_open` was called as a side-effect of the library initialization. If we determine this then we can also add stderr as a redirection for asl and get all the good logging.

Reviewed By: zeyadsalloum

Differential Revision: D15938667

fbshipit-source-id: f4941595af455ade773dcbb0c48838b9bceb301f