Fix writing to stream after return of instruments
Summary:
Building on the parent diff, if there is any async logging happening after the call returns, there will be a segfault as the writer may (or should) be deallocated since the queue is writng after the return of the handler thread.

This writes and end of file into the block, which has the effect of preventing any further callbacks from firing.

This also ensures that there are only ever reads and writes happening on a singular queue. The only async usage of this is in the logging and therefore needs to be stopped with `consumeEndOfFile` after it is started. This is effectively cancellation.

This is a little messy for the time being, we probably need to do a better job of bridging asynchronous writing in a sane way. But for the time being this at least prevents and nasty races here.

Reviewed By: zeyadsalloum

Differential Revision: D16049891

fbshipit-source-id: 3c748ab491a0464859a197cb4e4cb01cbea809b9