dispatch_io_close *does* relinquish the file descriptor
Summary:
There was an erroneous implementation here that meant that it was possible to hang a socket shutdown indefinately.

This is because there isn't a guarentee that `dispatch_io_create`'s "Cleanup Handler" will get called if `dispatch_io_close` is called. This behaviour can appear to be different between sockets and regular file descriptors.

Reviewed By: zeyadsalloum

Differential Revision: D16050460

fbshipit-source-id: 8e2a133c6a4fcc8c59999a0c97f2611b40811e65