kill spawned companion when daemon dies
Summary:
`idb kill` doesn't clean up any spawned companion. it only kills the daemon which left the companion running on a random port as we give it port 0 when we start it. this puts idb in a state where if one tries to connect to the companion manualyy using `idb connect localhost 10882` it would fail as the companion is not running on 10882.
this diff kills any spawned companion when the daemon exits

Reviewed By: c-ryan747

Differential Revision: D16111161

fbshipit-source-id: 42846f7e5c8db54413e7f5cfe64cc221853f952c