Pass in absolute paths to file commands
Summary:
Currently when the user provides relative paths to `idb file push` these are passed through as relative paths all the way through to the companion(!)

Thus they'll only succeed if the cwd of the companion is the same as the client. Resolving the paths in the client to avoid this.

Reviewed By: zeyadsalloum

Differential Revision: D16541724

fbshipit-source-id: fdd64d586acaaa57086fe3aecf9263448f1f6690