add ability to custom Instruments timings for apps
Summary:
We can now customize the timings for the Instruments operation. For Wilde, we need a slighly larger timeout for the `Instruments Trace Completed` message to show up in a faulure case.

For example, in this Wilde [test](https://our.intern.facebook.com/intern/chronos/jobinstance/?jobinstanceid=5515701566&smc=chronos_gp_admin_client), Instruments did not start properly, but IDB thought it was launched properly. This resulted in the QPL marker missing error.

Differential Revision: D16567429

fbshipit-source-id: d0c0a5905f61ad7c959034010d415c037de86d81