fix spawning notifier
Summary:
list targets was not spawning a notifier as it's new implementation was not calling get_stub(). this is a problem if list-targets is called before any other command. but since list-targets is normally the first command to be called i added another call there.

i have also added clearing the local_targets_file on `idb kill` as a cleanup

Reviewed By: zbigniew-malinowski

Differential Revision: D16985438

fbshipit-source-id: 933dc9f2a35fa7a50a473c87959658a6075cbe46