write to stdout when initial state is done
Summary: there was a race condition where the first time `idb list-targets` is called it would return nothing. it's the cli was not waiting for the notifier to finish writing before it returns the results. the old notifier used to write to stdout when the initial state is done and so i added it here again. another diff incoming in the cli to wait till this is written out to stdout before returning

Reviewed By: lawrencelomax

Differential Revision: D17282673

fbshipit-source-id: 33b8fcce5c5851cb451ea5e9bbc2305efe3c0062