Copy simulator name
Summary:
Occasionally the notifier would crash because `FBiOSTargetStateUpdate` had a weak reference to the simulators name. When the simulator object got deallocated this would happen to point to another `FBiOSTargetStateUpdate` object, crashing when we try and turn this into json.

Fixing this in the same way we deal with the `udid` by copying it instead

Reviewed By: CarlaCosenza

Differential Revision: D17761133

fbshipit-source-id: 8d91ba50b36a64ecc6f75a6e323641580d47d69f