Fix running logic test on Xcode 11.1
Summary:
In Xcode 11.1, if we use fbxctest to run logic test (unit test), and if some test failed, then xctest will crash in this case with following callstack:

```
0   CoreFoundation                      0x00007fff23baa1ee __exceptionPreprocess + 350
1   libobjc.A.dylib                     0x00007fff50864b20 objc_exception_throw + 48
2   CoreFoundation                      0x00007fff23ba9f68 +[NSException raise:format:arguments:] + 88
3   Foundation                          0x00007fff25614de9 -[NSAssertionHandler handleFailureInMethod:object:file:lineNumber:description:] + 191
4   XCTest                              0x00000001002bfa27 __39+[XCTRunnerDaemonSession sharedSession]_block_invoke + 400
5   libdispatch.dylib                   0x00007fff516ad781 _dispatch_client_callout + 8
6   libdispatch.dylib                   0x00007fff516ae9ad _dispatch_once_callout + 20
7   XCTest                              0x00000001002bf895 +[XCTRunnerDaemonSession sharedSession] + 101
8   XCTest                              0x00000001002932bf -[XCUIDevice initLocalDeviceWithPlatform:] + 177
9   XCTest                              0x00000001002931f9 __25+[XCUIDevice localDevice]_block_invoke + 37
10  libdispatch.dylib                   0x00007fff516ad781 _dispatch_client_callout + 8
11  libdispatch.dylib                   0x00007fff516ae9ad _dispatch_once_callout + 20
12  XCTest                              0x00000001002931d1 +[XCUIDevice localDevice] + 45
13  XCTest                              0x00000001002b7057 __51-[XCTestCase(Failures) _recordActivityWithFailure:]_block_invoke + 209
```

The direct reason is in Xcode 11.1, XCTest.framework's  `+[XCTRunnerDaemonSession sharedSession]` change implementation to force connecting to testmanagerd service on simulator/device, and if it cannot make the connection, it will trigger an assertion failure and crash xctest.

see following for compare different implementation between Xcode 11.1 (left) and Xcode 10.3 (right):

{F217779168}

P.S. this only happens when some test failed because in Xcode 11 it try to take screenshot automatically after test failed, regardless of if it's logic/app test or UI test, and it requires getting a XCUIDevice object for taking screenshot. `XCUIDevice` will try to connect to `testmanagerd` service.

Reviewed By: zeyadsalloum

Differential Revision: D17950356

fbshipit-source-id: a734d5976e8c95d82b4e296b2f387298a79a6554