Update device screenshot to synchronize on main queue
Summary: We've found cases where the screenshot command doesn't tear down its AMDevice connection. That combined with a recent finding that the AMDevice connection times out after being idle for 60 seconds leads to the "Failed to start * service" that we see periodically. I was able to recreate this locally with some cleverly placed breakpoints. Moving everything to synchronize on the main queue seems to remove this error both locally and in prod, so will do that for now to reduce the impact of this.

Reviewed By: thegreatwallfb

Differential Revision: D17966041

fbshipit-source-id: ca9c0ce1cbe07043730e6554a50f1c2fae232ec2