Support override xcode developer path for fbxctest
Summary:
Add a way to use environment variable to override xcode developer path.
This is for supporting a use case that sometimes we need to build the app
using latest SDK, but run app/test on older version of installed xcode without
switch different xcode versions.

This also fix an issue with video recording. Different supported video recording
arguments of simctl depends on the version of a private framework: CoreSimulator.framework,
and it is installed together with Xcode, and once installed it won't swtich to older
version if you swtich to a older version of Xcode, so here we should properly check
the version of simctl to decide which set of argument to use for video recording.

Differential Revision: D18162278

fbshipit-source-id: 33cc0e53d6d852a07b230d8ac89566800d1929a6