Teardown FBAMDevice immediately after starting services
Summary:
FBAMDevice only needs to be open long enough to start a service and can be closed immediately after. See FBAMDevice.h for much longer explanation of why.

In order to accomplish this, we needed to add a new FBFuture replace operation that pops a future context off the stack and replaces it with another future context.

Reviewed By: lawrencelomax

Differential Revision: D18636417

fbshipit-source-id: 334bd75c6eb1899b32373380e248cf9231cc1f77