Remove FBDeltaUpdateManager+Instruments
Summary:
The delta manager is not used as-is, instead we're just using a single static method that can be moved elsewhere.

As a result of this change we can remvoe the FBDeltaUpdateManager+Instruments as it's no longer used in a streamed-ipc model since all the state is present in the handler thread.

Reviewed By: c-ryan747

Differential Revision: D19391199

fbshipit-source-id: 0f33a9f6f71a8d05bfa6ce71eaa6393aaa97b61d