Replace Delta Updates w/ Streamed XCTest Results
Summary:
Now that the rest of the core of the XCTest request system has been extracted to a class-cluster, we can now remove the delta update based approaches.

This only existed when we were using non-streamed IPC and this acted as a bridge between the old and new worlds. Keeping this state around makes little sense since we can just keep the state in the context of the handler thread, which is already multi-threaded but has all of the operation occuring synchronously w.r.t to the handler thread.

This also has positive consequences for latency. We will now flush results as soon as they happen, rather than polling for them. This is especially true for short-lived tests. It also potentially means that the results can themselves be re-modeled around each write containing a single result. There are also positive implications for logging. This pushes through a single method call that means that manipulation is far easier. Also, log messages are flushed as soon as they occur. This means that for any hanging tests the output will be provided. The notion of "completion" is now far more accurate, it reflects when both the reporting and the test operation have completeted, rather than just the latter.

The sole usage of the `FBConsumableXCTestReporter` was idb for the purpose of this bridging. This class has now been adapted so that it does streamed writing of this over GRPC, so the push-to-poll adaptation has now been removed

Reviewed By: c-ryan747

Differential Revision: D19390072

fbshipit-source-id: 804ebf0f8463c12d6fe8a809f638e137b34ca9dc