fix idb crash when no test results exist in the bundle
Summary:
idb will [crash](https://our.intern.facebook.com/intern/paste/P125209853/) when the expected `tests` field does not exist in the result bundle. [Here](https://pxl.cl/WSZt) is one example. `tests` field might not exist if no test methods are excuted (e.g., when idb could not find the xctest file).

This is the root cause of the uptick for the `IdbException: ('Connection closed')` error. In this diff, if `tests` does not exist, we log the error message and label the test as failed. With this change, in MobileLab, the trial would fail in case `tests` does not exist, but we get the result bundle back and the next trial will continue.

Reviewed By: nslingerland

Differential Revision: D19451127

fbshipit-source-id: 238f6f1976d58cb34fa3b9f2fbdc356809e58cc0