IdbClientBase -> IdbClient, IdbClient -> IdbManagementClient
Summary:
Management related commands are the "exception rather than the rule" when it comes to command creation.

Therefore the naming should reflect that and we can then begin to move CLI parsers to care more about the base commands rather than the management related ones. This prevents over-disclosure of the management commands to cli parsers that are unconcerned with them.

The indentation changes are due to `black` and are temporary. Future diffs will make the majority of commands avoid using the management client, which will alter the methods that are exposed to them. This change is purely a rename

Reviewed By: c-ryan747

Differential Revision: D19446734

fbshipit-source-id: 8078db8af53e0f6508144a76ef52e6bf129f928d