Split Management vs. non Management Comamands into Sibling Classes
Summary:
This will allow further refinement so that CLI parsers declare whether they are involved with management or not.

This will allow the construction of commands where a "direct companion" can be created.

I've also broken out the client obtaining and parser methods into functions so that they can be shared. Using a CM for the client is desirable so that we can also fix the resource leak problem for `CompanionCommand` derived commands.

Now we have the overwhelming majority (non-management related) commands using direct client. This will then allow us to change the way that we obtain the client entirely.

Reviewed By: c-ryan747

Differential Revision: D19446828

fbshipit-source-id: d5e8da83323afaba2dd34d2dc05dd36e06ea8e45